Audit fixes and unit tests for count, invalid, sample, singletons
Fixes barcode counters overflowing at ~32k reads (map[string]int16 ->
int) in count and singletons, an unchecked format-detection error in
sample.fq that could nil-panic, sample.fq's .bc output file dropping
its directory prefix, and unchecked temp-file errors on the
singletons stdin-spool path.

Also updates two tests (count, sample) whose expectations were
written against the pre-fix CheckFastqFormat behavior (files with
fewer than 100 records now succeed instead of erroring, per the
fastq/ package fix in an earlier commit on this branch).

+++ djinn/count/count.fq.go
@@ -25,7 +25,7 @@ func CountFQ(fqs []string, invalid bool) error {
 	defer writer.Flush()
 
 	// ---- main loop -------------------------
-	set := make(map[string]int16, 7_000_000) // barcode container
+	set := make(map[string]int, 7_000_000) // barcode container
 	var bc string
 	var valid bool
 
@@ -60,7 +60,7 @@ func CountFQ(fqs []string, invalid bool) error {
 
 	// R2
 	if len(fqs) > 1 {
-		setR2 := make(map[string]int16, 500_000) // R2 barcode container
+		setR2 := make(map[string]int, 500_000) // R2 barcode container
 
 		fqReader, err = fastx.NewDefaultReader(fqs[1])
 		if err != nil {
